[#18] Provide slf4j 2.0 compatibility
* Implemented service loader
* All tests pass
* TODO: check service loader implementation
