+++ deps.edn
@@ -1,13 +1,14 @@
 {:paths
  ["src/clojure"
+  "resources"
   "target/classes"]
 
  :deps
  {org.clojure/clojure {:mvn/version "1.11.1"}
   org.clojure/data.json {:mvn/version "2.4.0"}
-  org.slf4j/slf4j-api {:mvn/version "1.7.36"}
-  org.slf4j/jul-to-slf4j {:mvn/version "1.7.36"}
-  org.slf4j/jcl-over-slf4j {:mvn/version "1.7.36"}
+  org.slf4j/slf4j-api {:mvn/version "2.0.3"}
+  org.slf4j/jul-to-slf4j {:mvn/version "2.0.3"}
+  org.slf4j/jcl-over-slf4j {:mvn/version "2.0.3"}
   io.aviso/pretty {:mvn/version "1.1.1"}
   aero/aero {:mvn/version "1.1.6"}}
 
@@ -41,7 +42,7 @@
 
   :build
   {:deps {org.clojure/clojure {:mvn/version "1.11.1"}
-          org.slf4j/slf4j-api {:mvn/version "1.7.36"}
+          org.slf4j/slf4j-api {:mvn/version "2.0.3"}
           org.clojure/tools.build {:mvn/version "0.8.4"}
           slipset/deps-deploy {:mvn/version "0.2.0"}}
    :ns-default build}}}
