+++ resources/META-INF/services/org.slf4j.spi.SLF4JServiceProvider
@@ -1,0 +1,1 @@
+dialog.logger.DialogServiceProvider
